+++ temp.h
@@ -1,7 +1,6 @@
 #ifndef __temp_H__
 #define __temp_H__
-#include "stdint.h"
 
-int Tempsensor_read(void);
+uint16_t Tempsensor_read(void);
 
 #endif // __temp_H__
